Topology iteration type changed

+++ src/main/java/Model/Topology.java
@@ -37,7 +37,7 @@ public class Topology {
     /**
      * Jumlah iterasi pada saat training model
      */
-    private double numIterations;
+    private int numIterations;
     /**
      * True if termination condition based on number of iterations
      * False if termination condition based on epoch error
@@ -46,7 +46,7 @@ public class Topology {
     /**
      * Do training until epoch error under this value
      */
-    private int epochErrorThreshold;
+    private double epochErrorThreshold;
     /**
      * True if termination condition based on epoch error
      * False if termination condition based on number of iterations
@@ -121,7 +121,7 @@ public class Topology {
         return numIterations;
     }
 
-    public void setNumIterations(double numIterations) {
+    public void setNumIterations(int numIterations) {
         this.numIterations = numIterations;
     }
